Added pagination

+++ web-api/src/main/java/com/app/model/PageResponse.java
@@ -2,40 +2,40 @@ package com.app.model;
 
 public class PageResponse extends BaseResponse {
 
-    private int total=0;
-    private int pageSize=0;
-    private int currentPage=0;
-    private int totalPages=0;
+    private Long total=0L;
+    private Long pageSize=0L;
+    private Long currentPage=0L;
+    private Long totalPages=0L;
     private String pagingMsg="";
     public boolean moreItemExist=false;
 
     // ==== Getters and Setters ====
-    public int getTotal() {return total;}
-    public void setTotal(int total) {
+    public Long getTotal() {return total;}
+    public void setTotal(Long total) {
         this.total = total;
         this.pageSize = total;
-        this.currentPage = 1;
-        this.totalPages = 1 ;
+        this.currentPage = 1L;
+        this.totalPages = 1L;
     }
 
-    public int getPageSize() {return pageSize;}
-    public void setPageSize(int pageSize) {this.pageSize = pageSize;}
+    public Long getPageSize() {return pageSize;}
+    public void setPageSize(Long pageSize) {this.pageSize = pageSize;}
 
-    public int getCurrentPage() {return currentPage;}
-    public void setCurrentPage(int currentPage) {this.currentPage = currentPage;}
+    public Long getCurrentPage() {return currentPage;}
+    public void setCurrentPage(Long currentPage) {this.currentPage = currentPage;}
 
-    public int getTotalPages() {return totalPages;}
-    public void setTotalPages(int totalPages) {this.totalPages = totalPages;}
+    public Long getTotalPages() {return totalPages;}
+    public void setTotalPages(Long totalPages) {this.totalPages = totalPages;}
 
     public String getPagingMsg() {return pagingMsg;}
     public void setPagingMsg(String pagingMsg) {this.pagingMsg = pagingMsg;}
 
-    public void setPageStats(int total, int pageSize, int currentPage, String pagingMsg){
+    public void setPageStats(Long total, Long pageSize, Long currentPage, String pagingMsg){
         if (pageSize>0){
             this.total = total;
             this.pageSize = pageSize;
             this.currentPage = currentPage;
-            this.totalPages = (int)(total/pageSize) +  ( (total % pageSize==0) ?0:1) ;
+            this.totalPages = (Long)(total/pageSize) +  ( (total % pageSize==0) ?0:1) ;
             this.pagingMsg = pagingMsg;
         }
     }
